# openapi.yaml
...
openapi: 3.1.0
info:
  title: openagentix API
  version: 0.1.0
  description: "Control node API of openagentix: agents, events, runs, approvals, audit, costs and RBAC."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Session or API token (`oax_...`)
    runToken:
      type: http
      scheme: bearer
      description: Signed run token (`oaxrt....`) of a worker node
    webhookSignature:
      type: apiKey
      in: header
      name: x-oax-signature
      description: HMAC-SHA256 signature
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: stable error code
        message:
          type: string
        details: {}
paths:
...
  /v1/users:
    get:
      summary: List users
      tags:
        - users
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        tenantId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        email:
                          type: string
                        displayName:
                          type: string
                        source:
                          type: string
                        globalRoles:
                          type: array
                          items:
                            type: string
                        disabled:
                          type: boolean
                        teams:
                          type: array
                          items:
                            type: object
                            properties:
                              teamId:
                                type: string
                                format: uuid
                                pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                              role:
                                type: string
                            required:
                              - teamId
                              - role
                            additionalProperties: false
                        createdAt:
                          type: string
                          description: ISO 8601 timestamp
                        lastLoginAt:
                          anyOf:
                            - type: string
                              description: ISO 8601 timestamp
                            - type: "null"
                      required:
                        - id
                        - tenantId
                        - email
                        - displayName
                        - source
                        - globalRoles
                        - disabled
                        - teams
                        - createdAt
                        - lastLoginAt
                      additionalProperties: false
                required:
                  - items
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a local user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  pattern: ^(?:[A-Za-z0-9_'+\-]+\.)*[A-Za-z0-9_'+\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 200
                password:
                  type: string
                  minLength: 12
                  maxLength: 200
                globalRoles:
                  default: []
                  type: array
                  items:
                    type: string
                    enum:
                      - admin
                      - agent-engineer
                      - integrator
                      - operator
                      - auditor
                      - viewer
              required:
                - email
                - displayName
                - password
      security:
        - bearer: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  tenantId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  email:
                    type: string
                  displayName:
                    type: string
                  source:
                    type: string
                  globalRoles:
                    type: array
                    items:
                      type: string
                  disabled:
                    type: boolean
                  teams:
                    type: array
                    items:
                      type: object
                      properties:
                        teamId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        role:
                          type: string
                      required:
                        - teamId
                        - role
                      additionalProperties: false
                  createdAt:
                    type: string
                    description: ISO 8601 timestamp
                  lastLoginAt:
                    anyOf:
                      - type: string
                        description: ISO 8601 timestamp
                      - type: "null"
                required:
                  - id
                  - tenantId
                  - email
                  - displayName
                  - source
                  - globalRoles
                  - disabled
                  - teams
                  - createdAt
                  - lastLoginAt
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  details: {}
                required:
                  - error
                  - message
                additionalProperties: false
  /v1/users/{id}:
    get:
      summary: Get a user
      tags:
        - users
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  tenantId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  email:
                    type: string
                  displayName:
                    type: string
                  source:
                    type: string
                  globalRoles:
                    type: array
                    items:
                      type: string
                  disabled:
                    type: boolean
                  teams:
                    type: array
                    items:
                      type: object
                      properties:
                        teamId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        role:
                          type: string
                      required:
                        - teamId
                        - role
                      additionalProperties: false
                  createdAt:
                    type: string
                    description: ISO 8601 timestamp
                  lastLoginAt:
                    anyOf:
                      - type: string
                        description: ISO 8601 timestamp
                      - type: "null"
                required:
                  - id
                  - tenantId
                  - email
                  - displayName
                  - source
                  - globalRoles
                  - disabled
                  - teams
                  - createdAt
                  - lastLoginAt
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update roles, name or disable a user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 200
                globalRoles:
                  type: array
                  items:
                    type: string
                    enum:
                      - admin
                      - agent-engineer
                      - integrator
                      - operator
                      - auditor
                      - viewer
                disabled:
                  type: boolean
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  tenantId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  email:
                    type: string
                  displayName:
                    type: string
                  source:
                    type: string
                  globalRoles:
                    type: array
                    items:
                      type: string
                  disabled:
                    type: boolean
                  teams:
                    type: array
                    items:
                      type: object
                      properties:
                        teamId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        role:
                          type: string
                      required:
                        - teamId
                        - role
                      additionalProperties: false
                  createdAt:
                    type: string
                    description: ISO 8601 timestamp
                  lastLoginAt:
                    anyOf:
                      - type: string
                        description: ISO 8601 timestamp
                      - type: "null"
                required:
                  - id
                  - tenantId
                  - email
                  - displayName
                  - source
                  - globalRoles
                  - disabled
                  - teams
                  - createdAt
                  - lastLoginAt
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/teams/{id}/members:
    get:
      summary: List the members of a team
      tags:
        - users
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        userId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        email:
                          type: string
                        displayName:
                          type: string
                        role:
                          type: string
                      required:
                        - userId
                        - email
                        - displayName
                        - role
                      additionalProperties: false
                required:
                  - items
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Replace the members (user + role) of a team
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                members:
                  maxItems: 1000
                  type: array
                  items:
                    type: object
                    properties:
                      userId:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      role:
                        type: string
                        enum:
                          - admin
                          - agent-engineer
                          - integrator
                          - operator
                          - auditor
                          - viewer
                    required:
                      - userId
                      - role
              required:
                - members
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/teams/{id}:
    patch:
      summary: Rename a team or change its monthly budget (null removes it)
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                monthlyBudgetUsd:
                  anyOf:
                    - type: number
                      exclusiveMinimum: 0
                    - type: "null"
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  slug:
                    type: string
                  name:
                    type: string
                  monthlyBudgetUsd:
                    type:
                      - number
                      - "null"
                  createdAt:
                    type: string
                    description: ISO 8601 timestamp
                required:
                  - id
                  - slug
                  - name
                  - monthlyBudgetUsd
                  - createdAt
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a team that owns no agents
      tags:
        - users
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "204":
          description: Default Response
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  details: {}
                required:
                  - error
                  - message
                additionalProperties: false
...
  /v1/tokens/{id}:
    delete:
      summary: Revoke an API token
      tags:
        - tokens
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-f]{16}$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: http://localhost:8080
    description: local development